Refactor job basics
