+++ src/tapis-ui/components/jobs/JobLauncher/steps/JobBasics.tsx
@@ -1,0 +1,94 @@
+import { useEffect } from 'react';
+import { Input } from 'reactstrap';
+import { FieldWrapper } from 'tapis-ui/_common';
+import { mapInnerRef } from 'tapis-ui/utils/forms';
+import { useFormContext } from 'react-hook-form';
+import { useWizard } from 'tapis-ui/_wrappers/Wizard';
+import { useJobLauncher } from '../JobLauncherProvider';
+import * as Jobs from '@tapis/tapis-typescript-jobs';
+
+type JobBasicsProps = {
+  appId: string;
+  appVersion: string;
+};
+
+export const JobBasics: React.FC<JobBasicsProps> = ({ appId, appVersion }) => {
+  const { register, reset, formState } = useFormContext<Jobs.ReqSubmitJob>();
+  const { goToStep } = useWizard();
+  const { errors } = formState;
+
+  // Reset to step 1 upon appId and appVersion change
+  useEffect(
+    () => {
+      reset({ 
+        name: `${appId}-${appVersion}-${new Date()
+          .toISOString()
+          .slice(0, -5)}`,
+        appId,
+        appVersion
+      });
+      goToStep && goToStep(1);
+    },
+    [ reset, goToStep, appId, appVersion ]
+  )
+  return (
+    <div>
+      <FieldWrapper
+        description="A name for this job"
+        label="Name"
+        required={true}
+        error={errors['name']}
+      >
+        <Input
+          bsSize="sm"
+          {...mapInnerRef(register('name', { required: 'Name is required' }))}
+        />
+      </FieldWrapper>
+      <FieldWrapper
+        description="The ID of the TAPIS application to run"
+        label="App ID"
+        required={true}
+        error={errors['appId']}
+      >
+        <Input
+          bsSize="sm"
+          data-testid="appId"
+          defaultValue={appId}
+          {...mapInnerRef(
+            register('appId', { required: 'App ID is required' })
+          )}
+        />
+      </FieldWrapper>
+      <FieldWrapper
+        description="The version of the application to run"
+        label="App Version"
+        required={true}
+        error={errors['appVersion']}
+      >
+        <Input
+          bsSize="sm"
+          defaultValue={appVersion}
+          {...mapInnerRef(
+            register('appVersion', { required: 'App version is required ' })
+          )}
+        />
+      </FieldWrapper>
+    </div>
+  );
+};
+
+export const JobBasicsSummary: React.FC = () => {
+  const { jobSubmission } = useJobLauncher();
+  const { name, appId, appVersion } = jobSubmission;
+  return (
+    <div>
+      {name && appId && appVersion
+        ? <div>
+            <div>{name}</div>
+            <div>{appId} v{appVersion}</div>
+          </div>
+        : <i>Incomplete</i>
+      }
+    </div>
+  )
+}
